First pass of refactoring glinit and makefiles; linux works
