+++ libswirl/utils/glinit/egl/egl.h
@@ -1,0 +1,7 @@
+#pragma once
+
+bool egl_MakeCurrent();
+void egl_GetCurrent();
+bool egl_Init(void* wind, void* disp);
+void egl_Swap();
+void egl_Term();
