+++ libswirl/utils/glinit/glx/glx.h
@@ -1,0 +1,7 @@
+#pragma once
+#include <GL/glx.h>
+
+bool glx_ChooseVisual(Display* display, long screen, XVisualInfo** visual, int* depth);
+bool glx_Init(void* wind, void* disp);
+void glx_Swap();
+void glx_Term();
