+++ libswirl/utils/glinit/sdl/sdl.h
@@ -1,0 +1,6 @@
+#pragma once
+
+void sdlgl_CreateWindow(bool gles);
+bool sdlgl_Init(void* wind, void* disp);
+void sdlgl_Swap();
+void sdlgl_Term();
