+++ libswirl/utils/glinit/wgl/wgl.h
@@ -1,0 +1,5 @@
+#pragma once
+
+bool wgl_Init(void* hwnd, void* hdc);
+void wgl_Swap();
+void wgl_Term();
